refactor(admin): share navbar styles across pages
