+++ web/sdk/admin/views/audit-logs/audit-logs.module.css
@@ -1,13 +1,3 @@
-.navbar {
-  padding: var(--rs-space-4) var(--rs-space-7);
-  border-bottom: 0.5px solid var(--rs-color-border-base-primary);
-  background: var(--rs-color-background-base-primary);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  min-height: var(--rs-space-12);
-}
-
 .table {
   height: auto;
 }
