+++ web/sdk/admin/views/organizations/list/list.module.css
@@ -1,13 +1,3 @@
-.navbar {
-  padding: var(--rs-space-4) var(--rs-space-7);
-  border-bottom: 0.5px solid var(--rs-color-border-base-primary);
-  background: var(--rs-color-background-base-primary);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  height: var(--rs-space-11);
-}
-
 .table {
   height: auto;
 }
